+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added slug, uri, id, uid, revisionNotes, revisionCreator, drafts to podcast and episode index pages
 - Added importerService
 - Added support for propagate-to and set-enabled-for-site for console resave
+- Added support for different podcast general settings per site
 - Now each podcast has two settings pages (general settings, episode settings)
 - Episode summary and description can be different now
 - Added basic support for pubDate on podcast RSS